fixed too many bugs

+++ src/middlewares/delete_messages.py
@@ -2,7 +2,7 @@ import logging
 
 from typing import Any, Callable, Dict, Awaitable
 
-from aiogram import BaseMiddleware, Bot
+from aiogram import BaseMiddleware
 from aiogram.exceptions import TelegramBadRequest
 from aiogram.types import TelegramObject, User, Message
 
@@ -16,19 +16,29 @@ class DeleteMessagesMiddleware(BaseMiddleware):
     ) -> Any:
 
         state_data = await data['state'].get_data()
+
         messages_ids = state_data.get('del_messages', [])
+        main_menu_message_id = state_data['main_menu_message_id']
+        # logging.info(f'{main_menu_message_id = }')
+        
         user: User = data["event_from_user"]
         bot = data['bot']
 
         for message_id in messages_ids:
-            try:
-                await bot.delete_message(chat_id=user.id, message_id=message_id)
-            except TelegramBadRequest:
-                pass
+            if not message_id == main_menu_message_id:
+                try:
+                    await bot.delete_message(chat_id=user.id, message_id=message_id)
+                    # logging.info(f'del message with id = {message_id}')
+                except TelegramBadRequest:
+                    pass
 
         result = await handler(event, data)
         if isinstance(event, Message):
-            if not (event.text and event.text.startswith('/')):
+            if not (
+                (event.text and event.text.startswith('/'))
+                or
+                event.message_id == main_menu_message_id
+            ):
                 try:
                     await bot.delete_message(chat_id=user.id, message_id=event.message_id)
                 except TelegramBadRequest:
